SAiDL-Summer-Assignment-2026/requirements.txt
torch
datasets
transformers
PyYAML
tqdm
wandb
pytest
torchvision
diffusers
accelerate
Pillow
clean-fid
scipy
